Improve: 프로필이미지, QR 코드 API 호출을 훅으로 추출
